Improve sampler objects handling
